rtapi:  Replace rev_code with serial
rtapi_data->rev_code and the REV_CODE macro were probably meant to be
a layout revision that was to be bumped after changes that break
compatibility.  However, it was repurposed to be a thread flavor
identifier instead.

This commit replaces rev_code with 'serial', and the REV_CODE macro
with RTAPI_SERIAL.  The RTAPI_SERIAL macro should be bumped whenever
changes that break compatibility are made.