libnml: Disable "strict aliasing"
This code violates the "strict aliasing" rule of C89, but nobody's likely to
fix this.  Forbid gcc from making optimizations enabled by the "strict
aliasing" rules (an option which exists because so much code was written before
optimizers began to systematically make optimizations that were possible in
light of the "strict aliasing" rule, like this code), and don't warn about
constructs that violate them.