libnml:  silence -Wstrict-aliasing warnings (fat patch)
The business of e.g. `*((u_long *) temp_buffer + 2)`, where
`temp_buffer` is a character buffer, breaks strict aliasing rules:

    libnml/buffer/tcpmem.cc:1842:69: warning: dereferencing type-punned pointer will break strict-aliasing rules [-Wstrict-aliasing]

The safe way to do this is make the buffer a union of a character
buffer and a uint32_t buffer.

No functions were used to manipulate these buffers (aside from the
'copy-and-paste' function), so this is a pretty massive change.