Refactor shm_common Cython bindings; make shm_common API generic
These shm_common Cython bindings are much more like real Python
objects.  Basic usage is as follows:

      key = 0x00baddad
      size = 512
      seg = shmcommon.new(key, size)
      # reattach
      seg = shmcommon.attach(key)
      shmcommon.exists(key)  # True
      seg.unlink()
      shmcommon.exists(key)  # False

Unit tests for the main functions are quite complete.

shm_common API generic:

This also introduces some changes in the shm_common code, as noted in
the key in hex format is appended to form the POSIX shm name.  This
makes the shm_common API generic, suitable for spinning out for use in
other projects.

This required removing the Machinekit-specific `instance` parameter
from the `shm_common_new` declaration, and thus fixing many calls
around HAL and RTAPI.  However, it was a minor change, since in all
cases the `key` parameter was passed in with the instance already
encoded.