shmdrv_api:  set key & size at __init__(); fix higher-level API
Setting the key and size attributes at __init__() allows segments to
be probed and unlinked without having to create or attach first. Duh.

This prompted refactoring in the higher-level API: The MKSHMSegment
class now may be used much like the lower-level SHMSegment class, but
subclasses like GlobalSegment may define default magic and size.

A lot of ugly trickiness was cleaned up as a result, esp. with the
`instance` attribute and funky descriptors.

Unit tests are also refactored.