WIP:  parameters
- Add errors
- Put value types in parameters & values
- Add iterators
  - Add subsection iterators
  - Add parameter iterators
  - Now we can walk the whole tree!
- Error on index gaps
- Add 'create' param to find_value()
-

FIXME: The Cython error handling is still a hack. Problem is when a
NULL value is returned, there's not a nice way to generate an
exception from Cython cdef functions. Best would be to do away with
the initial check and figure out how to deal with NULLs and
uninitialized strings where they should be dealt with.