Apply 'max velocity' to pure rotary motion
Add max angular velocity to UI parallel to max linear velocity, plumb
through task to motion, and in `tpGetMaxTargetVel()`, limit
pure-rotary motion.  UIs may then give user control of pure rotary
motion with a 'max velocity' slider.

- UI
  - `.ini` file:  `[TRAJ]MAX_ANGULAR_VELOCITY`
  - Python module
    - Add 2nd arg to `command.maxvel(vel_lin, vel_ang)`
    - Add `stat.max_angular_velocity` attribute
    - Update `axis.py` to compute angular velocity scale
      proportionately to linear velocity
  - `halui`: add `mav_*` pins in parallel to `mv_*` pins
  - `inihal`:  add `traj_max_angular_velocity` field
- NML
  - Add `velocity_angular` field to `EMC_TRAJ_SET_MAX_VELOCITY` message
  - Add `maxAngularVelocity` field to `EMC_TRAJ_STAT` message
- task
  - Add 2nd arg to `emcTrajSetMaxVelocity(maxvel_linear, maxvel_angular)`
- NML (previous commit)
- motion
  - Add `emcmot_config_t.limitVelAng` field
  - Add `emcmot_command_t.vela` field
- tp
  - Add 3rd arg to `tpSetVlimit()`

- Tests:
  - `motion/jogwheel`:  add ABC axes to test `halui`
  - `motion/max_velocity`:  add test for max velocity

Ugliness:

The [`.ini` docs][1] place `MAX_LINEAR_VELOCITY` and
`MAX_ANGULAR_VELOCITY` in the `[DISPLAY]` section; the former is the
maximum possible value on the max velocity slider.  The `MAX_VELOCITY`
param is in the `[TRAJ]` section, and is the set value on the slider.
This naming makes it hard to add the missing angular equivalent to
`[TRAJ]MAX_VELOCITY`.

[1]: http://linuxcnc.org/docs/2.7/html/config/ini-config.html